requirements.txt
genanki
load_dotenv
